minor dir refactoring
