+++ controllers/v1/file.js
@@ -1,12 +1,12 @@
 const {ErrorHandler, Utils} = require("../../helpers");
 const config = require("../../config");
-const {File} = require("../../repositories");
+const {File} = require("../../services");
 
 class FileController {
-    static create(req, res) {
+    create(req, res) {
         (async()=>{
             try{
-                let result = await File.Create(req.body.name, req.body.is_public, req.body.asset. req.user);
+                let result = await File.Create(req.body.name, req.body.is_public, req.body.asset, req.user);
                 let f = result.ops[0];
                 delete f.system_file_name;
                 return res.status(200).json({
@@ -24,7 +24,7 @@ class FileController {
         })();
     }
 
-    static get(req, res){
+    get(req, res){
         (async()=>{
             try{
                 let result = await File.Get(req.body.id);
@@ -44,23 +44,23 @@ class FileController {
         })();
     }
 
-    static getAll(req, res){
+    getAll(req, res){
         return res.status(200).json({
             message: "get files",
         });
     }
 
-    static update(req, res){
+    update(req, res){
         return res.status(200).json({
             message: "update files",
         });
     }
 
-    static delete(req, res){
+    delete(req, res){
         return res.status(200).json({
             message: "delete file",
         });
     }
 }
 
-module.exports = FileController;
+module.exports = new FileController();
