--- repositories/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-    User : require("./User"),
-    File : require("./File"),
-    Asset : require("./Asset")
-};
