+++ services/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    User : require("./user"),
+    File : require("./file"),
+    Asset : require("./asset")
+};
